Careful handling of executable paths on Windows
Check that user value, then default value of executable
path (i.e. gnuplot and feff) actually exists before attempting to run
them.  Also properly quote executable name in system call or pipe so
that things like spaces and parens are interpeted correctly.

This guards against a number of issues.  If Demeter is installed on
Windows in one location, then unistalled and reinstalled elsewhere,
the ini file will retain the incorrect value.  This also guards
against having an executable in a place with somethinng like
"system (x86)" or a unicode character in the path.

Also implement a correction from Brian Mattern to allow floats as
values in Hephaestus' ion chamber tool.