Modify Config.pm to handle Moose >= 2.0900
Moose now has a much more strict definition of Num in a type
constraint.  I would prefer not to add MooseX::Types::LaxNum as a new
dependence, so modified Config.pm to clean up numbers that pass
Scalar::Util's looks_like_number but not the new meaning of Num in
Moose.