Fixes #44
This incorporates two of the three changes in PR #44.

The most important change in this commit is to have Larch.pm end with a
flag indicating whether the Larch server was correctly started.  This is
essential so that Demeter.pm knows to revert to using ifeffit in the
case of larch_server not being available.

This also sidesteps the problem @newville reported with Athena's about dialog.